Model_Deployment/requirements.txt
tensorflow-cpu
numpy
flask
flask_cors
Pillow

